--- handlers/Movies.js
@@ -1,1 +1,0 @@
-'use strict';
