+++ handlers/errors.js
@@ -1,0 +1,8 @@
+'use strict';
+
+function handleErrors(err, response) {
+  console.log(err);
+  response.status(500).send(`Internal Error`);
+}
+
+module.exports = handleErrors;
